FE, BE(evluate_table, Seeder)

+++ boardinghouse_backend/database/migrations/2023_09_24_135019_create_evaluates_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('boarding_house_id');
-            $table->integer('ratings');
+            $table->double('ratings');
             $table->string('content')->nullable();
 
             $table->timestamp('created_at')->default(DB::raw('CURRENT_TIMESTAMP'));
